Add ensure shape config as a method, rename oci black box tests
This commit adds a ensureShapeConfig method to make sure we are setting
the correct shape config before launching an OCI instance.
Add white box unit tests on provider/oci and therefore rename all the
current tests to *_integration_test.go.
