+++ provider/oci/common_integration_test.go
