+++ provider/oci/images_integration_test.go
