+++ provider/oci/instance_integration_test.go
