+++ provider/oci/provider_integration_test.go
