+++ provider/oci/storage_integration_test.go
